requirements.txt
google-api-core==2.11.1
google-api-python-client==1.8.2
google-auth==2.22.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-secret-manager==2.15.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
pandas==1.0.5
pyarrow==12.0.1